requirements.txt
Flask
numpy
scikit-learn
tensorflow
pillow
werkzeug
huggingface-hub
gdown
gunicorn
gTTS
playsound
pandas